[add] Add Drm fd support (#501)
* [add] Add "--drm-fd" option to flutter-pi.

This allows to start flutter-pi on a specific DRM fd, instead of letting the tool locking the DRM master.
This can be useful if we have several screens and we want to display the flutter-pi content on a specific one, while not locking others.

This works well with the drm-lease-manager project (https://github.com/AGLExport/drm-lease-manager/tree/master)

* [update] Update README.md

---------

Co-authored-by: Renaud Barrau <renaud.barrau@madic.com>